+++ xtask/src/main.rs
@@ -1,7 +1,9 @@
+use is_executable::is_executable;
 use std::path::PathBuf;
 use std::process::Command;
 use std::{collections::HashMap, path::Path};
 use structopt::StructOpt;
+use strum::IntoEnumIterator;
 
 trait CommandExt: Sized {
     fn mode(&mut self, mode: BuildMode) -> &mut Self;
@@ -46,10 +48,21 @@ impl BuildMode {
     }
 }
 
-#[derive(Clone, Copy, PartialEq, Eq, Hash, strum_macros::EnumString, strum_macros::Display)]
+#[derive(
+    Clone,
+    Copy,
+    PartialEq,
+    Eq,
+    Hash,
+    strum_macros::EnumString,
+    strum_macros::Display,
+    strum_macros::EnumIter,
+)]
 enum Frontend {
     #[strum(to_string = "XIM")]
     Xim,
+    #[strum(to_string = "WAYLAND")]
+    Wayland,
     #[strum(to_string = "QT5")]
     Qt5,
     #[strum(to_string = "QT6")]
@@ -62,6 +75,15 @@ enum Frontend {
     Gtk4,
 }
 
+impl Frontend {
+    pub fn is_cmake(self) -> bool {
+        match self {
+            Frontend::Xim | Frontend::Wayland => false,
+            _ => true,
+        }
+    }
+}
+
 #[derive(StructOpt)]
 enum TaskCommand {
     Test,
@@ -69,7 +91,10 @@ enum TaskCommand {
     Build {
         #[structopt(long, parse(try_from_str), default_value = "Release")]
         mode: BuildMode,
-        #[structopt(parse(try_from_str))]
+        #[structopt(
+            parse(try_from_str),
+            about = "Select frontend availiable list: [XIM, WAYLAND, QT5, QT6, GTK2, GTK3, GTK4]"
+        )]
         frontends: Vec<Frontend>,
     },
     Install {
@@ -119,39 +144,36 @@ impl TaskCommand {
                 let out_path = get_build_path().join("out");
 
                 install(
-                    true,
                     out_path.join("kime-xim"),
                     target_path.join("usr/bin/kime-xim"),
                 );
                 install(
-                    true,
+                    out_path.join("kime-wayland"),
+                    target_path.join("usr/bin/kime-wayland"),
+                );
+                install(
                     out_path.join("libkime-gtk2.so"),
                     target_path.join("usr/lib/gtk-2.0/2.10.0/immodules/im-kime.so"),
                 );
                 install(
-                    true,
                     out_path.join("libkime-gtk3.so"),
                     target_path.join("usr/lib/gtk-3.0/3.0.0/immodules/im-kime.so"),
                 );
                 install(
-                    true,
                     out_path.join("libkime-gtk4.so"),
                     target_path.join("usr/lib/gtk-4.0/4.0.0/immodules/libkime-gtk4.so"),
                 );
-                install(true, out_path.join("libkime-qt5.so"), target_path.join("usr/lib/qt/plugins/platforminputcontexts/libkimeplatforminputcontextplugin.so"));
-                install(true, out_path.join("libkime-qt6.so"), target_path.join("usr/lib/qt6/plugins/platforminputcontexts/libkimeplatforminputcontextplugin.so"));
+                install(out_path.join("libkime-qt5.so"), target_path.join("usr/lib/qt/plugins/platforminputcontexts/libkimeplatforminputcontextplugin.so"));
+                install(out_path.join("libkime-qt6.so"), target_path.join("usr/lib/qt6/plugins/platforminputcontexts/libkimeplatforminputcontextplugin.so"));
                 install(
-                    true,
                     out_path.join("libkime_engine.so"),
                     target_path.join("usr/lib/libkime_engine.so"),
                 );
                 install(
-                    false,
                     out_path.join("kime_engine.h"),
                     target_path.join("usr/include/kime_engine.h"),
                 );
                 install(
-                    false,
                     out_path.join("config.yaml"),
                     target_path.join("etc/kime/config.yaml"),
                 );
@@ -167,17 +189,14 @@ impl TaskCommand {
             }
             TaskCommand::Clean => {}
             TaskCommand::Build { frontends, mode } => {
-                let mut build_xim = false;
-                let mut cmake_flags = [
-                    (Frontend::Gtk2, false),
-                    (Frontend::Gtk3, false),
-                    (Frontend::Gtk4, false),
-                    (Frontend::Qt5, false),
-                    (Frontend::Qt6, false),
-                ]
-                .iter()
-                .copied()
-                .collect::<HashMap<_, _>>();
+                let mut frontends = frontends
+                    .into_iter()
+                    .map(|f| (f, true))
+                    .collect::<HashMap<_, _>>();
+
+                for f in Frontend::iter() {
+                    frontends.entry(f).or_insert(false);
+                }
 
                 let src_path = get_src_path();
                 let build_path = get_build_path();
@@ -187,13 +206,6 @@ impl TaskCommand {
 
                 std::fs::create_dir_all(&out_path).expect("create out_path");
                 std::fs::create_dir_all(&cmake_out_path).expect("create cmake_out_path");
-
-                for frontend in frontends.iter() {
-                    match frontend {
-                        Frontend::Xim => build_xim = true,
-                        other => drop(cmake_flags.insert(*other, true)),
-                    }
-                }
 
                 // build engine core
                 build_core(mode);
@@ -206,7 +218,7 @@ impl TaskCommand {
                 )
                 .expect("Copy engine file");
 
-                if build_xim {
+                if frontends[&Frontend::Xim] {
                     Command::new("cargo")
                         .args(&["build", "--bin=kime-xim"])
                         .current_dir(get_src_path())
@@ -223,6 +235,23 @@ impl TaskCommand {
                     .expect("Copy xim file");
                 }
 
+                if frontends[&Frontend::Wayland] {
+                    Command::new("cargo")
+                        .args(&["build", "--bin=kime-wayland"])
+                        .current_dir(get_src_path())
+                        .mode(mode)
+                        .spawn()
+                        .expect("Spawn cargo")
+                        .wait()
+                        .expect("Run cargo");
+
+                    std::fs::copy(
+                        src_path.join(mode.cargo_target_dir()).join("kime-wayland"),
+                        out_path.join("kime-wayland"),
+                    )
+                    .expect("Copy xim file");
+                }
+
                 let mut cmake_command = Command::new("cmake");
 
                 cmake_command
@@ -231,7 +260,10 @@ impl TaskCommand {
                     .arg("-GNinja")
                     .arg(mode.cmake_build_type());
 
-                for (frontend, on) in cmake_flags.iter() {
+                for (frontend, on) in frontends.iter() {
+                    if !frontend.is_cmake() {
+                        continue;
+                    }
                     let flag = if *on { "ON" } else { "OFF" };
 
                     cmake_command.arg(format!("-DENABLE_{}={}", frontend, flag));
@@ -268,6 +300,10 @@ impl TaskCommand {
                     out_path.join("kime_engine.h"),
                 )
                 .expect("Copy default config file");
+
+                if mode.is_release() {
+                    strip_all(&out_path).ok();
+                }
             }
         }
     }
@@ -291,12 +327,36 @@ fn build_core(mode: BuildMode) {
         .expect("Run cargo");
 }
 
-fn install(exe: bool, src: PathBuf, target: PathBuf) {
+fn strip_all(dir: &Path) -> std::io::Result<()> {
+    for path in dir.read_dir()? {
+        let path = path?.path();
+
+        if !is_executable(&path) {
+            continue;
+        }
+
+        Command::new("strip")
+            .arg("-s")
+            .arg(path)
+            .spawn()
+            .expect("Spawn strip")
+            .wait()
+            .expect("Run strip");
+    }
+
+    Ok(())
+}
+
+fn install(src: PathBuf, target: PathBuf) {
     if src.exists() {
         println!("Install {} into {}", src.display(), target.display());
 
         Command::new("install")
-            .arg(if exe { "-Dsm755" } else { "-Dm644" })
+            .arg(if is_executable(&src) {
+                "-Dsm755"
+            } else {
+                "-Dm644"
+            })
             .arg(src)
             .arg("-T")
             .arg(target)