+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix NumLock bug again
 * Add GTK2, GTK4 immodule
 * Add Qt6 immodule
+* Add Wayland frontend
 
 # 0.8.1